// EIPs/erc4626/src/types/schema.ts
export class User {
  static entity = "User"
  static schema = {
    id: { type: "String", index: true },
    tokenBalance: "number",
    vaultBalance: "number",
    entryValue: "number",
    realizedEarnings: "number",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  }
}

export class Token {
  static entity = "Token"
  static schema = {
    id: { type: "String", index: true },
    name: "string",
    totalSupply: "number",
    multiplier: "number",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  }
}

export class Vault {
  static entity = "Vault"
  static schema = {
    id: { type: "String", index: true },
    name: "string",
    vaultAddress: "string",
    totalSupply: "number",
    totalHolding: "number",
    pricePerShare: "number",
    totalTokenEarnings: "number",
    minimumLock: "number",
    peripheryAddress: "string",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  }
}

export class dailyUserTrack {
  static entity = "dailyUserTrack"
  static schema = {
    id: { type: "String", index: true },
    userId: "String",
    dailyVaultBalance: "number",
    dailyEntryValue: "number",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  }
}

export class monthlyUserTrack {
  static entity = "monthlyUserTrack"
  static schema = {
    id: { type: "String", index: true },
    userId: "String",
    monthlyVaultBalance: "number",
    monthlyEntryValue: "number",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  }
}

export class yearlyUserTrack {
  static entity = "yearlyUserTrack"
  static schema = {
    id: { type: "String", index: true },
    userId: "String",
    yearlyVaultBalance: "number",
    yearlyEntryValue: "number",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  }
}

export class dailyVolume {
  static entity = "dailyVolume"
  static schema = {
    id: { type: "String", index: true },
    dailyVaultTotalSupply: "number",
    dailyPricePerShare: "number",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  }
}

export class monthlyVolume {
  static entity = "monthlyVolume"
  static schema = {
    id: { type: "String", index: true },
    monthlyVaultTotalSupply: "number",
    monthlyPricePerShare: "number",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  }
}

export class yearlyVolume {
  static entity = "yearlyVolume"
  static schema = {
    id: { type: "String", index: true },
    yearlyVaultTotalSupply: "number",
    yearlyPricePerShare: "number",
    entityId: { type: "String", index: true },
    blocknumber: { type: "Number", index: true },
    chainId: { type: "String", index: true },
    instanceId: { type: "String", index: true },
  }
}
// ...
export interface IUser extends Document {
  id: String
  tokenBalance: number
  vaultBalance: number
  entryValue: number
  realizedEarnings: number
  blocknumber: String
  entityId: String
  instanceId: String
  chainId: String
}

export interface IToken extends Document {
  id: String
  name: string
  totalSupply: number
  multiplier: number
  blocknumber: String
  entityId: String
  instanceId: String
  chainId: String
}

export interface IVault extends Document {
  id: String
  name: string
  vaultAddress: string
  totalSupply: number
  totalHolding: number
  pricePerShare: number
  totalTokenEarnings: number
  minimumLock: number
  peripheryAddress: string
  blocknumber: String
  entityId: String
  instanceId: String
  chainId: String
}